src/inference/requirements.txt
anyio==4.4.0
backoff==2.2.1
blinker==1.8.2
boto3==1.34.145
botocore==1.34.145
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
filelock==3.15.4
Flask==3.0.3
fsspec==2024.6.1
gql==3.5.0
graphql-core==3.2.3
huggingface-hub==0.24.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
MarkupSafe==2.1.5
mpmath==1.3.0
multidict==6.0.5
networkx==3.3
nltk==3.8.1
numpy==1.26.4
packaging==24.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.3
requests-toolbelt==1.0.0
s3transfer==0.10.2
safetensors==0.4.3
setuptools==69.5.1
six==1.16.0
sniffio==1.3.1
sympy==1.13.1
tokenizers==0.19.1
torch==2.3.1
tqdm==4.66.4
transformers==4.42.4
typing_extensions==4.12.2
urllib3==2.2.2
Werkzeug==3.0.3
wheel==0.43.0
yarl==1.9.4
